requirements.txt
Flask==2.3.3
Markdown==3.5
PyYAML==6.0.1
Werkzeug==2.3.7
Jinja2==3.1.2